Simplify/clarify history settings
Testing BASH_VERSINFO like this is really ugly, and not really worth the
convenience of infinite history in the upcoming Bash 4.3. There's also
little advantage in me having such huge amounts of history in memory for
each instance of my interactive shell; I've never even approached a
million lines of history on any of my machines. This should be more than
enough.

Clarified the differences between the variables a bit too.