Stop versioning backgrounds
This was a dumb idea. They're by far the biggest files in the repository
and they're not interesting to diff; all they'll do is slow things down.

Given that I've been breaking a lot of links lately in moving away from
GitHub, I may as well take the opportunity to purge the objects too.