Rewrite of kvlt(6)
It turns out GNU sed doesn't support POSIX word boundaries
[[:<:]],[[:>:]]. This leaves me with no compatible way to denote a word
boundary. I shouldn't really be surprised.

I've worked around this by padding the start and end of each line with a
tilde, and then removing it again at the end of the script, which is
not great but will have to do. It's preferable to having two-three
versions of each of the word-boundary rules to support ^ and $
anchoring.

This is probably the weirdest way anyone has ever learned sed.