Use porcelain commands for Git prompt
Checking the output of `git status -z` works, but to be correctly
handled requires using a null delimiter for read. Because I want to port
this to OpenBSD pdksh (which doesn't have the -d option for read), this
seems to be a workable alternative.

It also enables me to remove the process substitution (another thing
pdksh doesn't support), and the array of flags.

I haven't yet tested this for speed.