Significant improvements to mi5(1df)
A clunkier and probably-slower but more accurate parser--won't stumble
over quoted instances of the mi5(1df) delimiter within inline macro
expansions. This removes one of the CAVEATS described in the manual
page.

Also allow specifying the quote and unquote strings and also the dnl
string in m4, mostly for completeness' sake; the manual page warns
against this as I think it's probably missing the point of mi5(1df) if
you're getting to that point.