Escape % signs in prompt command output
The Zsh analogue of fd2b6fd. This one matters a bit more, as e.g.
"%test" is a legal Git branch name.

A generic fix for this for Korn shell will be tough, as some of them
seem to use backslash escape sequences and others don't. None of the
prompt commands *should* emit backslashes, and it doesn't allow remote
execution, so I might just leave that.