Move syntax .vimrc config into subfile
Some refactoring is done here, because as noted in 5caa13c, my custom
colorscheme is implemented as a plugin to be loaded by Pathogen, and
hence isn't available into after it's done its work.

I've removed the :set background line for now until I'm sure it's
needed, because at the moment I'm not sure.