Restore dotfiles(7) manual as pre-built dist file
Commit 70fcb35 incorrectly built the dotfiles(7) manual without the
header required by the Pandoc converter. Rebuilt it properly by making a
script in a new directory "dist" which is to be run by the maintainer
whenever its source file README.md is updated. This should probably be
automated on my end with Git hooks.

The reason we don't include the Pandoc recipe for making this manual as
a target in the Makefile is to do with the heavy dependency of Pandoc,
for which packages are not available on some desirable operating
systems, as arranged in a8ab2cf.

Also added the new install-man target as one of the default subtargets
of `install`.