Load Vim Pathogen with :runtime
Per this suggestion from the `vim-pathogen` FAQ:

<https://github.com/tpope/vim-pathogen#faq>
>>Can I put pathogen.vim in a submodule like all my other plugins?
>
>Sure, stick it under `~/.vim/bundle`, and prepend the following to
>your vimrc:
>
>   runtime bundle/vim-pathogen/autoload/pathogen.vim

This method avoids using symbolic links, which is desirable in general,
and also removes the need for the `install-vim-pathogen` dependency of
the `install-vim-plugin` target, since this is now done in Vim
configuration.

This also takes away another of the steps required for setting up the
Vim configuration on Windows.