Test whether to load sahara colorscheme
This is in preparation for config in vim/config/syntax.vim that will do
a more comprehensive job of applying heuristics to figure out if the
background is light or dark and hence what colours should be loaded for
the appropriate scheme.

The test for the GUI or 256 colours is repeated in the colorscheme code
itself, but I think that's OK given that sahara.vim is distributed
separately and others probably wouldn't use the kind of guards
introduced in this commit.