Switch on COLORFGBG to get background lightness
Now that this environment variable is kept and updated in tmux after
54553ae, we should be able to either configure terminals or explicitly
set it during startup if we want to use lighter terminals.

I'm much more comfortable with this than simply hardcoding it in the
configuration.

This doesn't solve the problem of carrying the environment variable over
an SSH session, however, but I'm not really sure there's a solution to
that besides configuring sshd(8) itself to accept these variables in
transit.