Move netrw.vim "after" script to plain config
There's no particular need to set these options after netrw has loaded;
they can be set before it's loaded, and the plugin will still observe
them.

This empties the vim/after/plugin directory. It doesn't need to be
removed from the Makefile as there are no references to it; it was
installed by a glob cp(1).