Rename Makefile target for Vim plugin bundles
It's misleading to label this target as installing plugins when what it
actually does at a directory level is install the Vim plugin submodules
into ~/.vim/bundle for loading by Pathogen.

This also allows scope for an `install-vim-plugins` target to actually
install into ~/.vim/bundle, if I do need that at some point.