Backport StripTrailingWhitespace to pre-for Vim
Use a slightly more verbose but more compatible `:while` loop to
accommodate very old versions of Vim that do not have `:for`.

Vim 6.2 gives the following terminal output when the `:for` version is
run:

    Error detected while processing /home/tom/.vim/config/whitespace.vim:
    line   13:
    E193: :endfunction not inside a function

However, it accepts this new version with no complaint, and the function
seems to work properly.