Show Vim option values after leader-key toggle
Using the very weird syntax:

    :setlocal option! option?

We can toggle the option and print its new value in the same line. For
the line breaking options, we only show the value of the 'linebreak'
option, because otherwise we print three lines of messages, which
requires an <Enter> press, even if we put all three `option?` calls on
one :setlocal line.