Revert "Drastically simplify `lint` scripts"
I forgot that the `lint` tools here need to check the *built* files, and
that that's the reason the `perlcritic` check against the source .pl
file was failing.

While it's still true that it would be preferable to test the files
found in a deterministic order, this branch's attempt to address that
issue is pretty much nonsense and can be abandoned.

This reverts commit 196155499c04b2c2050302e6575f1bcbbed052f1.