Apply stable check and lint methods to games shell
This applies the same stable approach to testing the actual built
games that are shebannged with #!/bin/sh as has been applied to the
shell scripts in the `check-bin` and `lint-bin` targets.

There are no GNU Bash games in these directories, so the latter block of
code from the `bin` analogues to check or lint those is not needed.

The same applies here; this is not as complete a checking or linting of
the games directory as it could be; ideally we would check the sed(1)
and awk(1) scripts too.