Tidy and correct linting for all three shells
Reduce each one to target only the dotfiles specifically for that shell,
as opposed to previously where for example the `check-sh` target was
checking shell shims in for `mpd` and `plenv`.

I'm still not completely sure that's the right approach, but it's at
least less conceptually muddy than what we had before.

Notably, the check and lint for Korn shell includes a single POSIX shell
script file in its `shrc.d` subdirectory, so that check is executed
separately.