Add dfv(1df) version-printing tool
This is mostly just for fun, but could be handy later on when I'm
playing with distributed or automated deployments of tagged and verified
releases.

Like a few of the other shell scripts, this is built by abusing my
mi5(1df) wrapper to get static details baked into the shell script that
are only known at runtime.