Move Vim big file options config into plugin
Created targets install-vim-doc and install-vim-plugin with accompanying
subdirectories of "vim".

Added a very short summary of what the plugin does to bigfile.txt.

I intend to spin off at least a couple of the blocks of my Vim
configuration that are starting to coalesce into distinct plugins unto
themselves, and will place the files in these directories.