Require eval feature for netrw opts assignment
Tiny builds of Vim that exclude the eval feature throw errors at the
:let commands in this file:

    $ vim.tiny
    Error detected while processing /home/tom/.vim/config/netrw.vim:
    line    2:
    E319: Sorry, the command is not available in this version: let g:netrw_banner = 0
    line    5:
    E319: Sorry, the command is not available in this version: let g:netrw_silent = 1
    line    8:
    E319: Sorry, the command is not available in this version: let g:netrw_liststyle = 3
    line   11:
    E319: Sorry, the command is not available in this version: let g:netrw_list_hide = '^\.$,^tags$'
    Press ENTER or type command to continue

This code was not being run on such builds before commit 538b71c,
because it was in an "after" directory for the netrw plugin and would
have been skipped.

Wrapping a check for has('eval') around this whole file fixes the
problem. netrw.vim won't run without this feature anyway.