Add tidy mapping for HTML
This mapping mirrors the one for Perl that passes the content of the
buffer through a program to tidy it (i.e. not merely check but actively
change it).

The tidy(1) option chosen here, -quiet, is the bare minimum to make this
invocation useful. We would never want the boilerplate it otherwise
emits to be in the buffer after a call. Everything else should be
applied in a configuration file, which I'll do in a separate feature.