Add ~/.tidyrc and accompanying Makefile target
This target also installs a short shell script in ~/.profile.d to set
and export the HTML_TIDY environment variable that defines the path to
the configuration file. tidy(1) seems to need this to be explicitly set
with a default build, as far as I can tell.

This pairs nicely with the settings in vim/ftplugin/html.vim.