Spin 'fo' toggle out into new flag toggler plugin
This is an experimental new plugin that provides a command to toggle
individual single-character flags in an option with a value of a set of
such flags, in my case 'formatoptions'.

A fair bit of evil eval()ing via :execute here, but I've tried to
control it with some strict patern matching.