Fix oii(1df) so it works as a pipe
I realised I could make this work by recording a single byte in the
temporary file with dd(1) and then emitting that and then the rest of
the input with cat(1) if the file ended up with a byte in it.

This lets me remove the CAVEATS section from the manual, as it no longer
applies.