Remove --batch option and wrapper for gpg(1)
This was originally added to cut the decryption boilerplate, which no
longer seems to be an issue; I think that --quiet may be correctly
blocking it now. Even without this, it caused more problems than it
solved when gpg(1) genuinely did need user interaction from me, for
example for --update-trustdb.