Make first ax(1df) arg safer, warn on second arg
The format in the first argument does not need to be evaluated, so it
can be passed in a simple awk variable.

The second argument is evaluated, by design, so code injection is
trivial. It's probably a good idea to warn users about this explicitly.

    $ ax '0);system("cat /etc/passwd")'

Make the whole thing a little terser, too, with the awk program
construction, variable assignment, and invocation all on one line.