Tweak 'cpoptions' C flag instead of 'nocompatible'
In order for the configuration to be successfully loaded, the only
option in the vi 'cpoptions' settings for 'compatible' is "C". From
:help 'cpoptions':

> C       Do not concatenate sourced lines that start with a backslash.
> See line-continuation.

With this flag removed from 'cpoptions' if 'compatible' does happen to
be set, the configuration parses just fine, and then we can put it back
at the end if we need to.

This is a less aggressive approach than just turning off 'compatible'
entirely if it happens to be set, whether because the user wanted it
that way before loading the configuration or because Vim was started as
ex(1).

My plugins and ftplugins are all conditional on 'compatible' not being
set, anyway.