Remove 'mac' from 'fileformats'
It seems unlikely that I'll ever edit a MacOS encoded file in my
lifetime on the Unix and Windows systems to which these dotfiles are
deployed, and when 'compatible' is set, the default empty value for this
option breaks everything with a bunch of ^J characters in every
god-fearing file. Not worth the trouble.