Remove vim-lion plugin for now
It doesn't have guards for old Vim, compatibility settings, or
repeat-reload checks, which is not very good, and means it spits errors
on old Vims about newer constructs like :finally, where all the other
plugins are well-behaved.

I was going to replace it with vim-easy-align, but that doesn't seem to
have a version or compatibility guard either, though it does have a
repeat-reload check which means I could probably shoehorn in a version
check before loading it, but even that seems a bit gross.

So, I might just leave lining things up nicely to the various tidy
scripts. Let's see how much I miss it.