Use script variable to keep showbreak setting
The 'showbreak' option is global and has no local value--that is, it's
not global-local--so `setlocal` actually just changes the global value,
and hence we can't "change back" using `setlocal showbreak<`. Instead,
keep the "normal" value in a script variable so that we can reliably
switch back to it.

This surprised me, as I had thought it was working, but I can't find
even an older Vim that behaves the way I expected it too. I suppose I
must just not have changed back often enough to actually notice.