Junk detect_background.vim and thereby autoload
It's too complicated and confusing, and doesn't do enough to justify
wrecking Vim's own logic for doing this sort of thing. Better to just
say `:set background=dark` and be done with it.

This is the only one of my inline plugins with an `autoload` file, so we
can get rid of that, too.

Not worth packaging/publishing to www.vim.org.