Add 'sessionoptions' settings
This disables saving global and local options and mappings in the
sessions. This is more trouble than it's worth for me, because sessions
don't include buffer-local variables, which means that in addition to
restoring the state of options and mappings, plugins and filetype
plugins are reloaded in the absence of b:loaded_* guards.