Scrap auto_cache_dirs plugin for a new approach
Trying this instead; creating the relevant directories at install time.
Also putting all three under one "cache" subdir of ~/.vim or ~/vimfiles.
I'm not sure this is actually better, but I do like that the relevant
options are now set in the .vimrc rather than in a plugin, so I guess
we'll see.