Add a dillorc file
I'm using this browser on an old ThinkPad. Firefox runs, but it's slow,
and lately I'm spending a lot of time on sites that don't require
JavaScript or anything like that. Let's see how this goes.

The `save_dir` doesn't accept tildes in it as a shorthand for the home
directory, so I've had to template it with mi5->m4.