Correct RHS of (un)important mail ftplugin map
This resolves the issue raised in f1b6f3b, where the mappings did not
work if the cursor was on the first line.

I had the colon `:` that starts the `ex` command and the `<C-U>` that
clears any existing command around the wrong way. Because CTRL-U in
normal mode scrolls up, it fails if we're already on the top line, and
the rest of the mapping looks to be skipped.

There don't seem to be any other instances of the same error that I can
find.