Remove vim-tiny workaround for 'undodir' setting
Commit 575f00d changed the setting for 'undodir' and similar path-based
settings to set the values conditionally based on the operating system
being used, while still setting them sensibly for "tiny" builds of Vim
on Unix, which skips all :if blocks.

This isn't necessary for 'undodir', because a tiny build of Vim won't
have the persistent_undo feature required for the option to exist at
all, so we can make this particular setting a little less awkward.

The 'backupdir' and 'directory' settings, however, need to keep their
existing structure to remain interoperable.