Add textarea filetype
This custom filetype is just a hook to hang functionality for
TextEditorAnywhere documents under Windows.  Most of the time, it's me
writing email, or comments that benefit from email formatting, so I've
configured <LocalLeader>f to switch to the mail filetype.  This
switching could probably benefit from a plugin, actually.