Add 'indent' and 'eol' flags to 'backspace'
It's annoying not being able to remove the indent in e.g. heredocs, and
I feel that being able to do so is still philosophically consistent,
because the spaces were inserted during that insert operation.  Newlines
were, too.

What I don't want to do, at least not just yet, is add 'start', because
I should be using normal mode commands for such deletions.