Drastically simplify $MYVIM handling
It was a fun exercise, but it's a pretty perverse situation to set up.
People generally don't put commas in their runtime directory names, and
it's not too out of order simply to tell them not to do that if they do.

This removes the need for two small utility functions.