Create Vim cache dir for new 'spellfile' location
Commit c0bab95, released in v5.31.0, moved the location of the
'spellfile' from the default of the "spell" subdir in the path in the
first element of &runtimedir to a custom path of $MYVIM/cache/spell.
The applicable code includes an attempt to create that directory should
it not already exist, but it's appropriate to create it at install time
anyway.